# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Talent Bridge API",
    "version": "1.0.0",
    "description": "API documentation for the Talent Bridge platform",
    "contact": {
      "name": "Talent Bridge Team",
      "email": "support@talentbridge.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "Employer": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Employer unique identifier",
            "example": "64f890abc123def456789013"
          },
          "userId": {
            "type": "string",
            "description": "Associated user ID",
            "example": "64f890abc123def456789013"
          },
          "companyName": {
            "type": "string",
            "description": "Company name",
            "example": "TechCorp Solutions"
          },
          "companyDescription": {
            "type": "string",
            "description": "Company description",
            "example": "Leading software development company specializing in web and mobile applications"
          },
          "industry": {
            "type": "string",
            "description": "Company industry",
            "example": "Information Technology"
          },
          "companySize": {
            "type": "string",
            "enum": [
              "1-10",
              "11-50",
              "51-200",
              "201-500",
              "501-1000",
              "1000+"
            ],
            "description": "Company size range",
            "example": "51-200"
          },
          "website": {
            "type": "string",
            "format": "url",
            "description": "Company website URL",
            "example": "https://techcorp.com"
          },
          "address": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string",
                "example": "456 Business Avenue"
              },
              "city": {
                "type": "string",
                "example": "Karachi"
              },
              "state": {
                "type": "string",
                "example": "Sindh"
              },
              "country": {
                "type": "string",
                "example": "Pakistan"
              },
              "postalCode": {
                "type": "string",
                "example": "75600"
              }
            }
          },
          "establishedYear": {
            "type": "integer",
            "description": "Year company was established",
            "example": 2015
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "pending",
              "verified",
              "rejected"
            ],
            "description": "Company verification status",
            "example": "verified"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Profile creation timestamp",
            "example": "2025-08-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Profile last update timestamp",
            "example": "2025-08-19T14:20:00.000Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Always false for error responses",
            "example": false
          },
          "statusCode": {
            "type": "integer",
            "description": "HTTP error status code",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "Error message describing what went wrong",
            "example": "Validation error or operation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Detailed error messages (optional)",
            "example": [
              "Email is required",
              "Password must be at least 6 characters"
            ]
          }
        }
      },
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "description": "Current page limit",
            "example": 20
          },
          "total": {
            "type": "integer",
            "description": "Total number of items",
            "example": 150
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages",
            "example": 8
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether there are more results available"
          },
          "nextCursor": {
            "type": "string",
            "description": "Cursor for the next page"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/employers": {
      "post": {
        "summary": "Create a new company profile",
        "description": "Create a new employer/company profile (Only available to users with employer role)",
        "tags": [
          "Employers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "companyName",
                  "companySize",
                  "industry"
                ],
                "properties": {
                  "companyName": {
                    "type": "string",
                    "description": "Company or organization name",
                    "example": "TechCorp Solutions"
                  },
                  "companyDescription": {
                    "type": "string",
                    "description": "Brief description of the company",
                    "example": "Leading software development company specializing in web and mobile applications"
                  },
                  "companySize": {
                    "type": "string",
                    "enum": [
                      "1-10",
                      "11-50",
                      "51-200",
                      "201-500",
                      "501-1000",
                      "1000+"
                    ],
                    "description": "Company size range",
                    "example": "51-200"
                  },
                  "industry": {
                    "type": "string",
                    "description": "Industry sector",
                    "example": "Information Technology"
                  },
                  "website": {
                    "type": "string",
                    "format": "url",
                    "description": "Company website URL",
                    "example": "https://techcorp.com"
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "street": {
                        "type": "string",
                        "example": "456 Business Avenue"
                      },
                      "city": {
                        "type": "string",
                        "example": "Karachi"
                      },
                      "state": {
                        "type": "string",
                        "example": "Sindh"
                      },
                      "country": {
                        "type": "string",
                        "example": "Pakistan"
                      },
                      "postalCode": {
                        "type": "string",
                        "example": "75600"
                      }
                    }
                  },
                  "establishedYear": {
                    "type": "integer",
                    "minimum": 1800,
                    "maximum": 2025,
                    "description": "Year company was established",
                    "example": 2015
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employer profile created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "employer": {
                          "$ref": "#/components/schemas/Employer"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Employer profile created successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only users with employer role can create company profiles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Company profile already exists for this user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all companies with pagination and filtering",
        "description": "Retrieve a paginated list of verified company profiles with optional filtering",
        "tags": [
          "Employers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination",
            "example": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            },
            "description": "Number of companies per page (max 50)",
            "example": 20
          },
          {
            "in": "query",
            "name": "industry",
            "schema": {
              "type": "string"
            },
            "description": "Filter by industry sector",
            "example": "Information Technology"
          },
          {
            "in": "query",
            "name": "companySize",
            "schema": {
              "type": "string",
              "enum": [
                "1-10",
                "11-50",
                "51-200",
                "201-500",
                "501-1000",
                "1000+"
              ]
            },
            "description": "Filter by company size",
            "example": "51-200"
          },
          {
            "in": "query",
            "name": "location",
            "schema": {
              "type": "string"
            },
            "description": "Filter by company location (city)",
            "example": "Karachi"
          }
        ],
        "responses": {
          "200": {
            "description": "Companies retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "companies": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Employer"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationInfo"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Companies retrieved successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No companies found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/employers/{id}": {
      "get": {
        "summary": "Get a company by ID",
        "tags": [
          "Employers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Company fetched successfully"
          },
          "404": {
            "description": "Company not found"
          }
        }
      },
      "put": {
        "summary": "Update company details",
        "tags": [
          "Employers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "companySize": {
                    "type": "string"
                  },
                  "industry": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company profile updated successfully"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Company not found"
          }
        }
      },
      "delete": {
        "summary": "Delete company profile and related jobs",
        "tags": [
          "Employers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Company profile and related jobs deleted successfully"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Company not found"
          }
        }
      }
    },
...
  }
}